refactor: split test modules that crossed the readability limit
trajectory_outbox.rs and store/ots.rs both grew past the 500-line
production-file gate as the JCS trajectory capture and OTS persistence
landed. Both move their test suites to sibling *_test.rs files — the
same pattern used for turn.rs, atif/mod.rs, and conformance/walk.rs
earlier in this branch.

No behavior change: the moved code is unmodified.

PROD_FILES_GT500: 86 → 84 (matches baseline)
